requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pyjwt==2.8.1
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
alembic==1.13.1
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
